--- a/libautoscoper/src/Tracker.cpp
+++ b/libautoscoper/src/Tracker.cpp
@@ -409,7 +409,7 @@ void Tracker::optimize(int frame, int dFrame, int repeats, int opt_method, unsig
     if (optimization_method == 0) // 0: PSO + Downhill Simplex, 1: Downhill Simplex
     {
       // PSO Algorithm
-      double xyzypr_manip[6] = { 0 };
+
       //int gBest = 0;
       //int gBestTest = 1000;
       int stall_iter = 0;
@@ -419,61 +419,19 @@ void Tracker::optimize(int frame, int dFrame, int repeats, int opt_method, unsig
       unsigned int MAX_EPOCHS = max_iter;
       unsigned int MAX_STALL = max_stall_iter;
 
-      float positions[NUM_OF_PARTICLES*NUM_OF_DIMENSIONS];
-      float velocities[NUM_OF_PARTICLES*NUM_OF_DIMENSIONS];
-      float pBests[NUM_OF_PARTICLES*NUM_OF_DIMENSIONS];
-      float gBest[NUM_OF_DIMENSIONS];
-
-      srand((unsigned)time(NULL));
-
-      for (int i = 0; i < NUM_OF_PARTICLES*NUM_OF_DIMENSIONS; i++)
-      {
-        if (i >= NUM_OF_DIMENSIONS) {
-          positions[i] = getRandom(START_RANGE_MIN, START_RANGE_MAX);
-        }
-        // First point will be the initial position
-        else {
-          //if (j > 1)
-          //{
-            positions[i] = (float)0.0;
-          //}
-          //else
-          //{
-          //  positions[i] = (float)0.2; // Noise for refinements
-          //}
-        }
-        pBests[i] = positions[i];
-        velocities[i] = 0;
-      }
-
-      for (int i = 0; i < NUM_OF_DIMENSIONS; i++)
-      {
-        gBest[i] = pBests[i];
-      }
-
       clock_t cpu_begin = clock();
-
-      pso(positions, velocities, pBests, gBest, MAX_EPOCHS, MAX_STALL);
-      //cuda_pso(positions, velocities, pBests, gBest, MAX_EPOCHS);
-
+      Particle gBest = pso(START_RANGE_MIN, START_RANGE_MAX, MAX_EPOCHS, MAX_STALL);
       clock_t cpu_end = clock();
 
       printf("Time elapsed:%10.3lf s\n", (double)(cpu_end - cpu_begin) / CLOCKS_PER_SEC);
 
-      std::cout << "Pose change from initial position: ";
-      for (int i = 0; i < NUM_OF_DIMENSIONS; i++)
-      {
-        if (i == NUM_OF_DIMENSIONS - 1)
-        {
-          std::cout << gBest[i] << std::endl;
-        }
-        else {
-          std::cout << gBest[i] << ",";
-        }
-        xyzypr_manip[i] = gBest[i];
-      }
+      using ::operator<<; // Access the stream operator from the global namespace
+      std::cout << "Pose change from initial position: " << gBest.Position << std::endl;
 
-      printf("Minimum NCC from PSO = %f\n", host_fitness_function(gBest));
+      printf("Minimum NCC from PSO = %f\n", gBest.NCC);
+
+      double xyzypr_manip[NUM_OF_DIMENSIONS] = { 0 };
+      std::copy(gBest.Position.begin(), gBest.Position.begin() + NUM_OF_DIMENSIONS, xyzypr_manip);
 
       manip = CoordFrame::from_xyzAxis_angle(xyzypr_manip);
       // PSO End
@@ -494,15 +452,6 @@ void Tracker::optimize(int frame, int dFrame, int repeats, int opt_method, unsig
       trial_.getYawCurve(-1)->insert(trial_.frame, xyzypr[3]);
       trial_.getPitchCurve(-1)->insert(trial_.frame, xyzypr[4]);
       trial_.getRollCurve(-1)->insert(trial_.frame, xyzypr[5]);
-
-
-      // Get Current Pose
-      double xyzypr[6] = { (*trial_.getXCurve(-1))(trial_.frame),
-        (*trial_.getYCurve(-1))(trial_.frame),
-        (*trial_.getZCurve(-1))(trial_.frame),
-        (*trial_.getYawCurve(-1))(trial_.frame),
-        (*trial_.getPitchCurve(-1))(trial_.frame),
-        (*trial_.getRollCurve(-1))(trial_.frame) };
 
       // DOWNHILL SIMPLEX
       // Generate the 7 vertices that form the initial simplex. Because