:warning: This pull request was created to include the changes originally included through this pull-request https://github.com/BrownBiomechanics/Autoscoper/pull/223

For the sake of having a clean history and considering the updated `main` branch has not been referenced externally, `main` was forced push to https://github.com/BrownBiomechanics/Autoscoper/commit/73eb359dcd70878c4b4c28471fd2684c08bd3014

----------

Refactor the Particle Swarm Optimization (PSO) implementation by introducing a specialized Particle data structure alongside a particle collection. This structure allows us to efficiently manage and track computed Normalized Cross-Correlation (NCC) values for each particle. As a result, we now evaluate each particle only once per iteration, as opposed to between 2 and 5 times in the worst-case scenario.

This optimization has led to a remarkable reduction in per-frame execution time on a typical workstation using the OpenCL back-end, dropping it from the range of 9 to 18 seconds to a much-improved 3 to 6 seconds.

In the previous implementation, a for loop iterated over all particles, involving multiple calls to the `host_fitness_function`:

* The first two calls determined if the current particle's NCC value was better than the "best" within the current iteration.
* The subsequent two calls were used to assess whether the current particle was better than the "global best".
* The final call was responsible for updating the global best value.

By introducing the Particle data structure and revising the workflow, we have streamlined the computation process, resulting in a substantial performance boost.